chore: remove exploratory multi-tag test
This test was exploring different API formats during development
of batch tag support (v4.1.0). Now that we know the correct format
({tags: [...]}) and it's implemented in production, this exploratory
test is no longer needed.

The working implementation is in TagManagementClient.ts and tested
in TagManagementClient.test.ts and TagManagementClient.integration.test.ts.